requirements.txt
adabelief_pytorch==0.2.1
albumentations==1.1.0
einops==0.3.2
entmax==1.0
matplotlib==3.5.1
numpy==1.17.3
pandas==1.1.4
Pillow==9.0.0
pytorch_lightning==1.5.8
scipy==1.5.4
setuptools==50.3.0
timm==0.4.12
torch==1.10.1
torchvision==0.11.2
tqdm==4.50.2
